Cambiar el Id de la tabla Log a BigInt Crear la tabla Track cambiar el ToolTip Actualizar a editar

+++ scripts/Full Database Creation Script.sql
@@ -1632,7 +1632,7 @@ CREATE SCHEMA [log]
 GO
 
 CREATE TABLE [log].[Logs](
-	[Id] [int] IDENTITY(1,1) NOT NULL,
+	[Id] [bigint] IDENTITY(1,1) NOT NULL,
 	[Message] [nvarchar](max) NULL,
 	[MessageTemplate] [nvarchar](max) NULL,
 	[Level] [nvarchar](max) NULL,
